Routing error fixed and added functionality
Fixed routing error that caused page load to throw errors.

Added more info about the contacts on the dashboard and added the ability to add a cellphone number and email to a contact.

+++ src/App/DisplayPeople.js
@@ -15,14 +15,19 @@ function DisplayPeople() {
 
     useEffect(() => {
         auth.onAuthStateChanged((user) => {
-            if (!user) navigatePages("/home");
+            if (!user) {
+                navigatePages("/");
+            } else {
+                setPeoplesContacts(JSON.parse(sessionStorage.getItem("contactAppContacts")));
+            }
         });
-        setPeoplesContacts(JSON.parse(sessionStorage.getItem("contactAppContacts")));
     }, []);
 
     const getContactsFromDB = async () => {
         try {
-            const data = await getDocs(collection(db, `${auth.currentUser.uid}contacts`));
+            const data = await getDocs(
+                collection(db, `${auth.currentUser.displayName}${auth.currentUser.uid}`)
+            );
             const contactPeople = data.docs.map((doc) => ({ ...doc.data(), id: doc.id }));
             setPeoplesContacts(contactPeople);
             sessionStorage.setItem("contactAppContacts", JSON.stringify(contactPeople));
@@ -37,7 +42,7 @@ function DisplayPeople() {
     const deleteItem = async (id) => {
         setDeletingContact([true, id]);
         try {
-            deleteDoc(doc(db, `${auth.currentUser.uid}contacts`, id));
+            deleteDoc(doc(db, `${auth.currentUser.displayName}${auth.currentUser.uid}`, id));
             getContactsFromDB();
             console.log("success");
         } catch (err) {
@@ -60,24 +65,29 @@ function DisplayPeople() {
         finalList = temporaryList;
     });
 
-    return (
-        <>
-            <Navbar page="contacts" />
-            <main>
-                <h2>Contacts</h2>
-                <SearchPeople handleSearch={setSearch} />
-                <div className="display_grid_wrapper grid">
-                    <h3 className="display_grid_item">ID</h3>
-                    <h3 className="display_grid_item">Info</h3>
-                </div>
-                <PeopleToRender
-                    deletePerson={deleteItem}
-                    people={finalList}
-                    deletingContact={deletingContact}
-                />
-            </main>
-        </>
-    );
+    if (auth.currentUser) {
+        return (
+            <>
+                <Navbar />
+                <main>
+                    <h2>Contacts</h2>
+                    <SearchPeople handleSearch={setSearch} />
+                    <p>
+                        Showing {finalList.length} of {peoplesContacts.length}
+                    </p>
+                    <div className="display_grid_wrapper grid">
+                        <h3 className="display_grid_item">Type</h3>
+                        <h3 className="display_grid_item">Info</h3>
+                    </div>
+                    <PeopleToRender
+                        deletePerson={deleteItem}
+                        people={finalList}
+                        deletingContact={deletingContact}
+                    />
+                </main>
+            </>
+        );
+    }
 }
 
 export default DisplayPeople;
